changed according to comment
Thank you for your detailed answer. Anyhow, I dont see how this is mapping requests to different controllers based on the port the request was received on What I need are separate controllers which are used for request on different ports. I.e. a request for "/x" incoming on port 8080 needs to be handled by a different (!) controller than the same request for "/x" being received on port 8081. I played around with the code provided by you, but I can't see how this is implementing the desired solution as all requests via all ports are handled by the same controller. Am I missing something?